fix: restore static typing health — 49 errors to 0
mypy was configured but never enforced. Baseline was 49 errors in 16 files; the
rehabilitation work pushed it to 53 before this pass. Now zero across all 173
source files, and CI runs `mypy --no-incremental .` over the whole repository
rather than a hand-picked subset.

Fixed by root cause, not suppression. The largest cluster was a single idiom:

  features/extractors/odds_movement.py — 12 errors from one line. The guard
    `if None in (a, b, c, d)` narrows nothing, so every later arithmetic use of
    four Optional floats was an error. Rewritten as an explicit `is None`
    disjunction, which narrows all four. 12 errors, one change.

One finding was a real latent bug rather than a typing nit:

  api/routes/tab_tracking.py — the /today endpoint read `rec.tier` and
    `rec.data_quality_ok`. Neither is a column on Recommendation and neither
    ever has been, so every call to that endpoint raised AttributeError. The
    path had no test coverage, which is why it survived. Confirmed by
    instantiating the model and reading the attributes. Now reported as null,
    with a comment saying the fields do not exist rather than pretending they
    do.

Other root causes: heterogeneous accumulator dicts annotated once at
declaration rather than inferred from their first branch; Optional columns
(`external_id`, `match_time`, `clv`) filtered or asserted at the point where the
code already knew they were set; `np.mean` narrowed to float; SQLAlchemy
Integer-backed booleans assigned bool rather than int; `self._model` declared
Optional-until-fit; a factory map typed `Callable[[], BaseModel]` instead of
inferring `object`.

Four new `# type: ignore` comments, all narrow, error-code-specific and
explained: two ModuleType stubs in a test, one heterogeneous test kwargs dict,
one Optional narrowed on the following line. No `ignore_errors`, no blanket
`Any`, no mass suppression. The pre-existing `ignore_missing_imports` for
third-party stubs is unchanged.

Canonical metrics verified unchanged after the whole pass — Brier, log loss,
accuracy, pooled and season-weighted ECE are byte-identical for all six models.
399 passed, 1 skipped. ruff clean.

+++ models/elo_baseline.py
@@ -163,7 +163,7 @@ class EloBaseline(BaseModel):
 
     def _season_regression(self) -> None:
         """Regress all team ratings towards the mean at the start of a new season."""
-        mean = np.mean(list(self.ratings.values())) if self.ratings else INITIAL_RATING
+        mean = float(np.mean(list(self.ratings.values()))) if self.ratings else INITIAL_RATING
         self.ratings = {
             team_id: rating + self.season_regression * (mean - rating)
             for team_id, rating in self.ratings.items()
